requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.11.3
python-dotenv==1.1.1
pydantic==2.11.3
pydantic-settings==2.8.1
